Time: 20 ms (61.04%), Space: 199.2 MB (56.42%) - LeetHub

+++ 3834-merge-adjacent-equal-elements/3834-merge-adjacent-equal-elements.cpp
@@ -12,7 +12,7 @@ public:
 
             while(stack.size()>=2 && stack[stack.size()-1] == stack[stack.size()-2])
             {
-                int val = stack.back();
+                long long val = stack.back();
                 stack.pop_back();
                 stack.back()+=val;
             }
